Improve dictionary subclassing
Use a little metaprogramming to make the code less repetitive, while still maintaining the goal of quacking like a dictionary. In the case of the `YadisManager`, this requires changing the API a little bit to avoid interfering with `dict` built-in methods (particularly `__len__` and `__iter__`), but these seem little-used so it seems safe from the tests.