Update dependencies, test actions
Newer versions of GitHub Actions want a `requirements.txt` file, so we now provide one, and have updated the `setup.py` and `dev-requirements.txt` for compatibility as well. While we're at it, we also need to tweak the GitHub Actions workflows to make things work correctly on older Pythons and on PyPy.